\name{bimodalGEA-package}
\alias{bimodalGEA-package}
\alias{bimodalGEA}
\docType{package}
\title{
  bimodality-based Gene Expression data analysis
}
\description{
  bimodality-based Gene Expression data analysis
}
\details{
  bimodality-based Gene Expression data analysis.
}
\author{
Antonio, Fabio Di Narzo

Maintainer: Antonio, Fabio Di Narzo <AntonioFabio.DiNarzo@isb-sib.ch>
}
\keyword{ package }
