\name{crossTable}
\alias{crossTable}
\title{cross-tabulate mixture-based expression levels with another factor
  variable}
\description{
  cross-tabulate mixture-based expression levels with another factor
  variable
}
\usage{
crossTable(x, fac)
}
\arguments{
  \item{x}{gene expression levels}
  \item{fac}{factor}
}
\keyword{ misc }
\keyword{ util }
