\name{mixTable}
\alias{mixTable}
\title{summarize a dataset by two levels of a continuous variable
}
\description{summarize a dataset by two levels of a continuous variable}
\usage{
mixTable(x, df)
}
\arguments{
  \item{x}{gene expression vector}
  \item{df}{dataframe to be summarized}
}
\author{
  Antonio, Fabio Di Narzo
}
\keyword{ misc }
\keyword{ util }
