\name{bimodalGEA-internal}
\title{Internal bimodalGEA objects}
\alias{.compose}
\alias{.formatMixParameters}
\alias{.plotMixParameters}
\alias{.mixDiscretize}
\alias{varBetween}
\alias{varTot}
\alias{varWithin}
\alias{delta}
\description{Internal bimodalGEA objects.}
\details{These are not to be called by the user.}
\keyword{internal}
